# openapi.yaml
openapi: 3.1.0
info:
  title: Business Management & POS System API
  version: 1.0.0
  summary: REST + JSON contract for the single-business web-based POS and operations platform.
  description: Authoritative machine-readable contract derived from API-Architecture-V1.0.md. Server is the only authority
    for business rules, accounting, and inventory. All derived balances (COGS, payment_state, AR/AP/CRL/SRec, cash_balance,
    inventory_value, moving_average_unit_cost, on_hand_quantity) are server-computed. Posted records are immutable; corrections
    go through lifecycle actions (cancel, return, refund). Cancellation, return, and refund are distinct events. All amounts
    use NUMERIC(15,2); quantities NUMERIC(15,4). Single currency (IDR). All timestamps are ISO 8601 with offset; server normalizes
    to UTC. Session-based auth (Bearer opaque tokens) per §2 of the API Architecture.
  contact:
    name: Business POS System API Team
  license:
    name: Proprietary
servers:
- url: https://api.example.com/api/v1
  description: Production
- url: https://staging-api.example.com/api/v1
  description: Staging
- url: http://localhost:8000/api/v1
  description: Local development
...
paths:
...
  /exports:
    post:
      operationId: createExport
      summary: Create an export job (PDF/XLSX). Rate-limited 5/hour/user.
      description: Create an export job (PDF/XLSX). Rate-limited 5/hour/user.
      tags:
      - Exports
      security:
      - bearerAuth: []
      x-required-capability: export.data
      parameters:
      - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /exports/{id}:
    parameters:
    - name: id
      in: path
      required: true
      description: Export job id (UUID)
      schema:
        type: integer
        format: uuid
    get:
      operationId: getExport
      summary: Get export job status and (when complete) download URL.
      description: Get export job status and (when complete) download URL.
      tags:
      - Exports
      security:
      - bearerAuth: []
      x-required-capability: export.data
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /exports/{id}/download:
    parameters:
    - name: id
      in: path
      required: true
      description: Export job id (UUID)
      schema:
        type: integer
        format: uuid
    get:
      summary: Download export file (pre-signed URL returned in 302 or streamed body).
      description: Returns a 302 redirect to the pre-signed download URL, or streams the file.
      operationId: downloadExport
      tags:
      - Exports
      security:
      - bearerAuth: []
      responses:
        '200':
          description: File binary.
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '302':
          description: Redirect to pre-signed URL.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Export expired or pruned.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Opaque
      description: 'Session-token auth per API Architecture §2. Server-issued opaque random tokens after credential exchange.
        256-bit random, base64url-encoded, stored hashed in the sessions table. Client does NOT parse them. Send `Authorization:
        Bearer <token>`. 401 = missing/expired/revoked. 403 = authenticated but missing capability.'
  parameters:
...
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: Required. See `IdempotencyKey`.
      schema:
        type: string
        format: uuid
        maxLength: 100
...
  responses:
    BadRequest:
      description: Validation failure (400).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorEnvelope'
    Unauthorized:
      description: Missing, invalid, or expired credentials (401).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Authenticated but missing required capability (403).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthorizationErrorEnvelope'
    NotFound:
      description: Resource not found (404).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: State conflict (409).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ConflictErrorEnvelope'
...
  schemas:
    ErrorEnvelope:
      type: object
      required:
      - error
      description: Canonical error envelope. Used for all 4xx/5xx responses.
      properties:
        error:
          type: object
          required:
          - code
          - message
          - request_id
          properties:
            code:
              type: string
              description: Stable, machine-readable snake_case code.
              example: not_found
            message:
              type: string
              description: Human-readable English.
            details:
              description: Optional, code-specific structure.
              oneOf:
              - type: object
                additionalProperties: true
              - type: array
                items:
                  type: object
                  additionalProperties: true
              - type: 'null'
            request_id:
              type: string
              format: uuid
              description: Echoes `X-Request-ID` header.
    ValidationErrorEnvelope:
      allOf:
      - $ref: '#/components/schemas/ErrorEnvelope'
      - type: object
        description: Returned for `validation_failed` (400).
        properties:
          error:
            type: object
            properties:
              code:
                const: validation_failed
              details:
                type: array
                items:
                  type: object
                  required:
                  - field
                  - code
                  - message
                  properties:
                    field:
                      type: string
                    code:
                      type: string
                    message:
                      type: string
    AuthorizationErrorEnvelope:
      allOf:
      - $ref: '#/components/schemas/ErrorEnvelope'
      - type: object
        description: Returned for `capability_required` (403).
        properties:
          error:
            type: object
            properties:
              code:
                const: capability_required
              details:
                type: object
                properties:
                  required:
                    type: array
                    items:
                      type: string
                    example:
                    - sale.cancel
    ConflictErrorEnvelope:
      allOf:
      - $ref: '#/components/schemas/ErrorEnvelope'
      - type: object
        description: Returned for state conflicts (409).
        properties:
          error:
            type: object
            properties:
              code:
                type: string
                enum:
                - insufficient_stock
                - negative_stock_disallowed
                - payment_exceeds_total
                - allocation_exceeds_payable
                - refund_exceeds_crl
                - repayment_exceeds_srec
                - cash_insufficient
                - return_quantity_exceeds_original
                - purchase_return_quantity_exceeds_original
                - lifecycle_state_invalid
                - already_cancelled
                - cannot_cancel_after_return
                - production_inputs_exceed_stock
                - production_finished_cost_invalid
                - cost_type_deactivated
                - product_deactivated
                - concurrent_modification
                - serialization_failure
                - idempotency_violation
                - delete_not_permitted
                - referenced_by_history
                - version_mismatch
              details:
                type: object
                additionalProperties: true
...
    ExportCreateRequest:
      type: object
      required:
      - report
      - format
      properties:
        report:
          type: string
          enum:
          - sales
          - purchases
          - inventory
          - inventory-movements
          - sales-returns
          - purchase-returns
          - production
          - manual-income
          - manual-expense
          - refunds
          - supplier-repayments
          - cash-flow
          - p-and-l
          - receivables
          - payables
          - supplier-receivables
          - customer-refund-liabilities
        format:
          type: string
          enum:
          - pdf
          - xlsx
        filters:
          type: object
          additionalProperties: true
    ExportJob:
      type: object
      required:
      - export_id
      - status
      - created_at
      properties:
        export_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - queued
          - processing
          - complete
          - failed
        download_url:
          type:
          - string
          - 'null'
          format: uri
        expires_at:
          type:
          - string
          - 'null'
          format: date-time
        error:
          type:
          - string
          - 'null'
        created_at:
          type: string
          format: date-time
...
security:
- bearerAuth: []